Fix invisible window issue and refactor window sizer (#1201)
* Refactor the window resizer class

This was difficult to follow with the class spread out over multiple mixin functions.
This rewrites it to be a class which subclasses the original class.

* Recover windows outside of the display area

If a window location was on a monitor which is no longer connected it would get moved there regardless making it difficult to recover.
This code checks if the window header is visible and recovers it automatically if it is not.

* Reformat